refactor(direct-control): extract traditions view source
Move the traditions view embedded source into packages/civ7-direct-control/src/play/progression/traditions.ts while keeping the public wrapper and command builder in the package facade. Keep the tiny probe helper local to avoid introducing a broad runtime-source helper module.

Record the DRA correction-channel fallback protocol in the support modularization workstream and mark the traditions source-owner step complete. This is source relocation with local fake-tuner package proof only, not runtime proof.

Validation: bun run --cwd packages/civ7-direct-control test -- progression-reads.test.ts; bun run --cwd packages/civ7-direct-control test; bun run --cwd packages/civ7-direct-control check; bun run --cwd packages/civ7-direct-control build; bun run openspec -- validate civ7-support-direct-control-modularization --strict; bun run resources:status; git diff --check.
